requirements.txt
biopython==1.84
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
tensorflow==2.19.0
keras==3.11.1
matplotlib==3.9.2
seaborn==0.13.2
streamlit==1.38.0